myDjangoProject/requirement.txt
Django==1.11
django-cors-headers==2.4.0
django-extensions==2.1.4
djangorestframework==3.9.0
mysqlclient==1.3.14
pytz==2018.9
six==1.12.0
